Make docs pretty
I've tried to make the docs look prettier for these singledispatch functions.
This will need a look from @falexwolf and @flying-sheep, though I may merge this PR before that.

I've disabled the error in the param_police, so that I could have type annotation in the docstrings.
It would be nice if all methods of generic functions were used for typing arguments.

Additionally, I'm not happy with the current ipython docs, since the signature only includes the arguments for the method that was decorated with `@singledistpatch`.