Rewrite _download
As noted in #1334, visium downloads were broken. Setting a header on downloads seems to fix them.

We can't use the solution suggested in #1334 since it modifies global state around `urllib`, which is asking for trouble.
This unfortunatley means most of the method had to be reimplemeted.
The new implementation is based on `urllib.requests.urlretrieve`, but with a modification to let us pass a header.

I've also removed the automatic `mkdir(..., parents=True)` from `_download` since typos could result in very messy filesystems.