Update _anndata.py
When using `seaborn==0.11`, the vertical orientation is ignored when
only `x` is specified. To orientate the violin plot vertically,
`seaborn.catplot` can be used instead of `seaborn.FacetGrid`.
The keyword arguments `"cut"` and `"inner"` cannot be passed to
`seaborn.catplot` or `seaborn.stripplot` and are thus moved to the
`else` clause.