Convert 'components' to 'dimensions' is sc.pl.embedding
Mostly this was to get rid of `_get_data_points`, which was a bit of a horrible function.

Now we get all values of the array, then for each plot index into it with the dimensions. This has removed a number of edge cases and special handling thoughout, making the dataflow a bit easier to follow.