Fixes #1097
sc.pl.scatter() is a wrapper for _scatter_obs(). It checks to make sure
the variable names the caller is requesting to plot exist in var and/or
obs, but does not take into account whether it should look in raw based
on the use_raw flag, as _scatter_obs() does. This leads to errors when a
user asks to plot variables that are in the raw but not the filtered
matrix of adata. This commit fixes that bug.