Update for cope with issue introduced in umap-learn 0.5.2 (#2028)
* Update for cope with issue introduced in umap-learn 0.5.2

As of umap-learn 0.5.2 an ``n_epochs`` value of 0 is now a valid value and None is used instead to indicate a value should be generated. To maintain forward compatability and backward compatibility with older umap-learn versions we instead make use of the epch setting criterion internal to umap-learn here directly. This should resolve issue #2026 and lmcinnes/umap#798

* Extra param that is needed before internal validate parameters call

* ... on the umap object

* Add release note

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>