Fix finding variables with sc.pl.scatter(..., use_raw=True) (#2027)
* Fixes #1097

sc.pl.scatter() is a wrapper for _scatter_obs(). It checks to make sure
the variable names the caller is requesting to plot exist in var and/or
obs, but does not take into account whether it should look in raw based
on the use_raw flag, as _scatter_obs() does. This leads to errors when a
user asks to plot variables that are in the raw but not the filtered
matrix of adata. This commit fixes that bug.

* Only use adata.raw.var for index, not keys

If `use_raw in [None, True]`, we want to look up gene names in
adata.raw.var, but we still want to look up var columns in adata.var.
Thus, this reverts the part of ec6ae95e084f37eaa4146645920ee61bf4dc9d34
that mistakenly tries to look up var_names in adata.raw.var.keys().

* Added tests for sc.pl.scatter(..., basis=None)

* Blacked

* Remove accidentally included test image

* Different tests for scatter successes and failures

As requested in code review, divide tests of `sc.pl.scatter()` with
`basis=None` into those in which the function should successfully make a
plot and those in which the function should raise a `ValueError`.

* Add test for scatterplots of raw with no basis

* Break scatter tests into per-var and per-obs

* Fix path typo

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>

* Plot all three first, then check

This allows having all the files for debugging in case the first comparison fails.

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>

* Use tmpdir instead of FIGS

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>

* Blacked

* Add release note

* Fix release note formatting

* Name change to release note

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>

Co-authored-by: Isaac Virshup <ivirshup@gmail.com>