n_pcs are no longer ignored for non (scanpy-internal) X_pca reductions
This change allows to only take the dimensions needed also for non X_pca representations. This can be useful when using harmony for example or storing different PCA embeddings in the same object. Then it is not needed to first store them as X_pca to be able to fully use the n_pcs parameter. It might make sense then to adapt the documentation as such as well.