n_pcs are no longer ignored for representations other than "X_pca" (#2179)
* n_pcs are no longer ignored for non (scanpy-internal) X_pca reductions

This change allows to only take the dimensions needed also for non X_pca representations. This can be useful when using harmony for example or storing different PCA embeddings in the same object. Then it is not needed to first store them as X_pca to be able to fully use the n_pcs parameter. It might make sense then to adapt the documentation as such as well.

* Added error if n_pcs > dimensions in representation

* Fixed typo in variable name

* Fixed styling

* using consistent quotation marks; all done from my side

* first idea for test; check in detail tomorrow

* Used wrong quotes before

* shortened lines

* Improved style

* improved style in new function

* trying to finish the test today after all

* Removing wrong toarray()

* Test changes as per pre-commit

* Function changes as per pre-commit

* Another try at implementing pre-commit changes

* Added release note

* Attempt to fix reference error

Co-authored-by: pmajev <paul-georg.majev@mpi-muenster.mpg.de>