Simplify embeddings a bit (#1538)
* Remove dependency on execution order for scatter plot tests

Previously `test_scatterplots[umap_with_edges]` would fail if it ran before `test_scatterplots[umap]`, since the `umap` case changed the color palette the anndata object used.
Now the AnnData object is copied for each test that uses it.

This is important because it means we don't have state spilling between the tests and can parallelize them.

* Convert 'components' to 'dimensions' is sc.pl.embedding

Mostly this was to get rid of `_get_data_points`, which was a bit of a horrible function.

Now we get all values of the array, then for each plot index into it with the dimensions. This has removed a number of edge cases and special handling thoughout, making the dataflow a bit easier to follow.

* Rearrange code in embedding to make a bit more sense

* Fix paga plots

* Implement broadcasting of dims and color

* Add simple test for broadcasting of dimensions and color

* Type dimensions

* add test for components and dimensions being interoperable

* Added docs for the dimension argument

* Undo deprecation of components kwarg.

I still want to eventually do that, but it shows up in many other parts of the codebase which I don't have time to replace at the moment. This includes, but is not limited to, the other pca plotting methods.

* Release note

* Fix type annotation for python 3.7